+++ js/paymill/paymentForm.js
@@ -10,6 +10,7 @@ var PAYMILL_ERROR_STRING = "";
 
 var eventFlag = false;
 
+
 function getPaymillCode()
 {
 	var methods = {
@@ -30,8 +31,7 @@ function getPaymillCode()
  */
 function debug(message)
 {
-	debug_state = pmQuery('.paymill-option-debug-' + getPaymillCode()).val();
-	if (debug_state === 1) {
+	if (pmQuery('.paymill-option-debug-' + getPaymillCode()).val() == '1') {
 		var displayName = "";
 		if (PAYMILL_PAYMENT_NAME === 'paymill_creditcard') {
 			displayName = 'Credit Card';
@@ -42,8 +42,9 @@ function debug(message)
 		if (PAYMILL_PAYMENT_NAME === 'Preparing Payment') {
 			displayName = 'Preparing Payment';
 		}
-
-		console.log("[" + displayName + "] " + message);
+		
+		console.log("[" + displayName + "]");
+		console.log(message);
 	}
 }
 
@@ -85,8 +86,7 @@ function paymillResponseHandler(error, result)
 		);
 
 		Object.extend(Validation.methods, nv);
-		debug(error.apierror);
-		debug(error.message);
+		debug(error);
 		debug("Paymill Response Handler triggered: Error.");
 	} else {
 		// Appending Token to form
@@ -150,20 +150,37 @@ function paymillSubmitForm()
 			break;
 		case "paymill_directdebit":
 			if (pmQuery('.paymill-info-fastCheckout-elv').val() === 'false') {
-				var valid = pmQuery('#paymill_directdebit_holdername').val() !== ''
-						 && paymill.validateAccountNumber(pmQuery('#paymill_directdebit_account').val())
-						 && paymill.validateBankCode(pmQuery('#paymill_directdebit_bankcode').val());
+				if (pmQuery('.paymill-info-sepa-elv').val() === 'false') {
+					var valid = pmQuery('#paymill_directdebit_holdername').val() !== ''
+							 && paymill.validateAccountNumber(pmQuery('#paymill_directdebit_account').val())
+							 && paymill.validateBankCode(pmQuery('#paymill_directdebit_bankcode').val());
 
-				if (!valid) {
-					return false;
+					if (!valid) {
+						return false;
+					}
+
+					debug("Generating Token");
+					paymill.createToken({
+						number: pmQuery('#paymill_directdebit_account').val(),
+						bank: pmQuery('#paymill_directdebit_bankcode').val(),
+						accountholder: pmQuery('#paymill_directdebit_holdername').val()
+					}, paymillResponseHandler);
+				} else {
+					var valid = pmQuery('#paymill_directdebit_holdername').val() !== ''
+							 && pmQuery('#paymill_directdebit_iban').val() !== ''
+							 && pmQuery('#paymill_directdebit_bic').val()  !== '';
+
+					if (!valid) {
+						return false;
+					}
+
+					debug("Generating Token");
+					paymill.createToken({
+						iban: pmQuery('#paymill_directdebit_iban').val(),
+						bic: pmQuery('#paymill_directdebit_bic').val(),
+						accountholder: pmQuery('#paymill_directdebit_holdername').val()
+					}, paymillResponseHandler);
 				}
-
-				debug("Generating Token");
-				paymill.createToken({
-					number: pmQuery('#paymill_directdebit_account').val(),
-					bank: pmQuery('#paymill_directdebit_bankcode').val(),
-					accountholder: pmQuery('#paymill_directdebit_holdername').val()
-				}, paymillResponseHandler);
 			}
 			break;
 	}
@@ -205,6 +222,22 @@ function unsetElvValidationRules()
 		'paymill-validate-dd-holdername': new Validator(
 			'paymill-validate-dd-holdername',
 			'',
+			function(v) {
+				return true;
+			},
+			''
+		),
+		'paymill-validate-dd-iban': new Validator(
+			'paymill-validate-dd-iban',
+			getValueIfExist('.paymill-payment-error-iban-elv'),
+			function(v) {
+				return true;
+			},
+			''
+		),
+		'paymill-validate-dd-bic': new Validator(
+			'paymill-validate-dd-bic',
+			getValueIfExist('.paymill-payment-error-bic-elv'),
 			function(v) {
 				return true;
 			},
@@ -285,6 +318,22 @@ function setElvValidationRules()
 		'paymill-validate-dd-holdername': new Validator(
 			'paymill-validate-dd-holdername',
 			getValueIfExist('.paymill-payment-error-holder-elv'),
+			function(v) {
+				return !(v === '');
+			},
+			''
+		),
+		'paymill-validate-dd-iban': new Validator(
+			'paymill-validate-dd-iban',
+			getValueIfExist('.paymill-payment-error-iban-elv'),
+			function(v) {
+				return !(v === '');
+			},
+			''
+		),
+		'paymill-validate-dd-bic': new Validator(
+			'paymill-validate-dd-bic',
+			getValueIfExist('.paymill-payment-error-bic-elv'),
 			function(v) {
 				return !(v === '');
 			},
@@ -378,6 +427,20 @@ function addPaymillEvents()
 	}
 	
 	if (!eventFlag) {
+		pmQuery('#paymill_directdebit_iban').keyup(function() {
+			var iban = pmQuery('#paymill_directdebit_iban').val();
+			if (!iban.match(/^DE/)) {
+				var newVal = "DE";
+				if (iban.match(/^.{2}(.*)/)) {
+					newVal += iban.match(/^.{2}(.*)/)[1];
+				}
+				pmQuery('#paymill_directdebit_iban').val(newVal);
+			}
+		});
+		
+		pmQuery('#paymill_directdebit_iban').trigger('keyup');
+
+		
 		pmQuery('#paymill_directdebit_holdername').live('focus', function() {
 			setElvValidationRules();
 			pmQuery('.paymill-info-fastCheckout-elv').val('false');
@@ -389,6 +452,16 @@ function addPaymillEvents()
 		});
 
 		pmQuery('#paymill_directdebit_bankcode').live('focus', function() {
+			setElvValidationRules();
+			pmQuery('.paymill-info-fastCheckout-elv').val('false');
+		});
+
+		pmQuery('#paymill_directdebit_iban').live('focus', function() {
+			setElvValidationRules();
+			pmQuery('.paymill-info-fastCheckout-elv').val('false');
+		});
+
+		pmQuery('#paymill_directdebit_bic').live('focus', function() {
 			setElvValidationRules();
 			pmQuery('.paymill-info-fastCheckout-elv').val('false');
 		});
@@ -439,6 +512,14 @@ function addPaymillEvents()
 		});
 
 		pmQuery('#paymill_directdebit_account').live('input', function() {
+			paymillSubmitForm();
+		});
+
+		pmQuery('#paymill_directdebit_iban').live('input', function() {
+			paymillSubmitForm();
+		});
+
+		pmQuery('#paymill_directdebit_bic').live('input', function() {
 			paymillSubmitForm();
 		});